Bugs and Minor UI Tweaks
BUGS:  When hand holds 2 aces BOTH are changed to value of 1.  TODO
added in Game class.  When playing as a user, it does not tell user when
they bust. TODO also in Game class. Redundancy with sayYouLose and
sayYouWin methods.  Has code to let player know these but is not used.
TWEAKS: Changed showResult method to show number of pushes and
reevaluated percentages on ALL games played not just ones that didn't
push.  Changed it to display percentage of games that pushed too.